+++ Binance.Net/Objects/BinanceAccountInfo.cs
@@ -10,19 +10,19 @@ namespace Binance.Net.Objects
         /// <summary>
         /// Commission percentage to pay when making trades
         /// </summary>
-        public double MakerCommission { get; set; }
+        public decimal MakerCommission { get; set; }
         /// <summary>
         /// Commission percentage to pay when taking trades
         /// </summary>
-        public double TakerCommission { get; set; }
+        public decimal TakerCommission { get; set; }
         /// <summary>
         /// Commission percentage to buy when buying
         /// </summary>
-        public double BuyerCommission { get; set; }
+        public decimal BuyerCommission { get; set; }
         /// <summary>
         /// Commission percentage to buy when selling
         /// </summary>
-        public double SellerCommission { get; set; }
+        public decimal SellerCommission { get; set; }
         /// <summary>
         /// Boolean indicating if this account can trade
         /// </summary>
@@ -53,14 +53,14 @@ namespace Binance.Net.Objects
         /// <summary>
         /// The amount that isn't locked in a trade
         /// </summary>
-        public double Free { get; set; }
+        public decimal Free { get; set; }
         /// <summary>
         /// The amount that is currently locked in a trade
         /// </summary>
-        public double Locked { get; set; }
+        public decimal Locked { get; set; }
         /// <summary>
         /// The total balance of this asset (Free + Locked)
         /// </summary>
-        public double Total => Free + Locked;
+        public decimal Total => Free + Locked;
     }
 }
